Moved the logOut function to blacklistedToken controller
